Added the variable analysis module that can be optionally executed before running k-induction.
Variable analysis moves variables to their most local scopes and removes variables that are unused. This reduces the state space, hopefully. For ECC and DSR, this clearly works (Removing around 10 to 20 percent of the variables, and moving 30 to 60 percent of the variables to the most local scope). Moving to the local scope means that for k-induction, the variable does not need to be havoc'ed.