Further work on integrating slicing.
For this, I had to change the typedef resolvation a little. It is now done recursively (for aggregates and their members) as well as by means of a fixed point iteration. This was needed due to the possibility of multiple "layers" of typedefs. Frama-C produced such code, which lead to some issues.