Typedef resolvation was restructured such that it does not work in-place on the AST anymore.
Seriously, how had this idea? It is way simpler to copy the declaration before resolving and return the copied declaration. This way, we still have access to the resolved variables, but leave the AST intact. Hence, we do not need to worry about transforming into valid C, which was problematic before, e.g. when creating multiple declarations of the same struct type.