Restructured the witness input a little bit.
Witness file location can now be given through the CLI - This needed to be done for the benchmarks. Also, I\'m not sure we need two separate witnesses for the base and induction call. Either the property holds (Then the induction call producs a witness) or it does not (Then the base call produces a witness), but it should not happen that both witnesses are written for the same input.